rename to hqpillow_retry
