+++ corehq/apps/hqpillow_retry/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import *
-from corehq.apps.pillow_retry.views import EditPillowError
+from corehq.apps.hqpillow_retry.views import EditPillowError
 
-urlpatterns = patterns('corehq.apps.pillow_retry.views',
+urlpatterns = patterns('corehq.apps.hqpillow_retry.views',
     url(r'^edit_errors/$', EditPillowError.as_view(), name=EditPillowError.urlname),
 
 )
