+++ urls.py
@@ -82,7 +82,7 @@ urlpatterns = patterns('',
     (r'^hq/admin/', include('corehq.apps.hqadmin.urls')),
     (r'^hq/reports/', include(report_urls)),
     (r'^hq/flags/', include('corehq.apps.toggle_ui.urls')),
-    (r'^hq/pillow_errors/', include('corehq.apps.pillow_retry.urls')),
+    (r'^hq/pillow_errors/', include('corehq.apps.hqpillow_retry.urls')),
     (r'^couchlog/', include('couchlog.urls')),
     (r'^formtranslate/', include('formtranslate.urls')),
     (r'^unicel/', include('corehq.apps.unicel.urls')),
